feat: KDEV-547 Add tests for StoreType.Network and StoreType.Auto
